summary
Last adds 27.12.2024

+++ src/main/java/com/example/pet_adopting/controllers/BreedController.java
@@ -4,6 +4,8 @@ import com.example.pet_adopting.entities.Breed;
 import com.example.pet_adopting.requests.CreateBreedRequest;
 import com.example.pet_adopting.requests.UpdateBreedRequest;
 import com.example.pet_adopting.services.BreedService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,27 +21,42 @@ public class BreedController {
     }
 
     @GetMapping
-    public List<Breed> getAllBreeds() {
-        return breedService.getAllBreeds();
+    public ResponseEntity<List<Breed>> getAllBreeds() {
+        List<Breed> breeds = breedService.getAllBreeds();
+        return ResponseEntity.ok(breeds);
     }
 
     @GetMapping("/{id}")
-    public Breed getBreedById(@PathVariable Long id) {
-        return breedService.getBreedById(id);
+    public ResponseEntity<Breed> getBreedById(@PathVariable Long id) {
+        Breed breed = breedService.getBreedById(id);
+        if (breed == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        return ResponseEntity.ok(breed);
     }
 
     @PostMapping
-    public Breed createBreed(@RequestBody CreateBreedRequest breed) {
-        return breedService.createBreed(breed);
+    public ResponseEntity<Breed> createBreed(@RequestBody CreateBreedRequest breedRequest) {
+        Breed createdBreed = breedService.createBreed(breedRequest);
+        return ResponseEntity.status(HttpStatus.CREATED).body(createdBreed);
     }
 
     @PutMapping("/{id}")
-    public Breed updateBreed(@PathVariable Long id, @RequestBody UpdateBreedRequest breed) {
-        return breedService.updateBreed(id, breed);
+    public ResponseEntity<Breed> updateBreed(@PathVariable Long id, @RequestBody UpdateBreedRequest breedRequest) {
+        Breed updatedBreed = breedService.updateBreed(id, breedRequest);
+        if (updatedBreed == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        return ResponseEntity.ok(updatedBreed);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteBreed(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteBreed(@PathVariable Long id) {
+        Breed breed = breedService.getBreedById(id);
+        if (breed == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
         breedService.deleteBreed(id);
+        return ResponseEntity.noContent().build();
     }
 }
